[TASK] Read max file upload size from php.ini configuration
This commit introduces a new utility class: "Configuration". The
purpose of this class is to provide details about the current
server/instance configuration, such as settings in file php.ini.

The first application of this utility class is to determine the
max file upload size and to populate this value in a numeric and
human-readable format in the Fluid templates/partials, so that the
JavaScript that validates the file uploads (attachments) uses a
dynamic, rather than a hard-coded value.