[FEATURE] Decouple editor and single post display from templates
By separating the input text area of forum posts (create and edit) from
Fluid templates, it becomes relatively easy to implement any editor and
text output handler, e.g. as a separate extension.

It should be noted, that the changes in this commit don't impact the
current functionality or visual appearance of the forum at all. The
main purpose of this update is to *enable* integrators and developers
to build their own solution.